Updates on trace

+++ src/delete-transaction.php
@@ -6,12 +6,12 @@
 
     if(array_key_exists("tranID", $_GET)) {
         $result['deleted'] = false;
-        $is_deleted = delete_transaction($_GET['tranID']);
-
-        if($is_deleted) {
-            $result['deleted'] = true;
-            save_trace("transaction-delete", $_SESSION['uid']);
-        } 
+        if(save_trace("transaction-delete", $_SESSION['uid'], $_GET['tranID'])) {
+            $is_deleted = delete_transaction($_GET['tranID']);
+            if($is_deleted) {
+                $result['deleted'] = true;
+            } 
+        }
     }
 
     echo json_encode($result);
